+++ src/main/java/org/kdb/inside/brains/view/chart/ChartTool.java
@@ -2,13 +2,15 @@ package org.kdb.inside.brains.view.chart;
 
 import com.intellij.openapi.actionSystem.ActionGroup;
 import org.jfree.chart.JFreeChart;
+import org.kdb.inside.brains.KdbType;
 
 public interface ChartTool {
-    void setChart(JFreeChart chart);
-
     boolean isEnabled();
 
     void setEnabled(boolean state);
+
+
+    void initialize(JFreeChart chart, KdbType domainType);
 
 
     default ActionGroup getPopupActions() {